+++ .github/workflows/primary.yml
@@ -23,7 +23,13 @@ jobs:
       DEVELOPER_DIR: /Applications/Xcode_12.app/Contents/Developer
     steps:
       - uses: actions/checkout@v2
-      - name: Build Package and Demos
-        run: sh ci.sh
+      - name: Build package
+        run: swift build
+      - name: Run unit tests
+        run: swift test
+      - name: Build Demos
+        run: sh ci/build_demos.sh
+      - name: Build xcframework
+        run: sh ci/xcframework.sh
       - name: Pod Lib Lint
         run: pod lib lint
