--- Sources/SVGColors.plist
@@ -1,302 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
-<plist version="1.0">
-<dict>
-	<key>aliceblue</key>
-	<string>#f0f8ff</string>
-	<key>antiquewhite</key>
-	<string>#faebd7</string>
-	<key>aqua</key>
-	<string>#00ffff</string>
-	<key>aquamarine</key>
-	<string>#7fffd4</string>
-	<key>azure</key>
-	<string>#f0ffff</string>
-	<key>beige</key>
-	<string>#f5f5dc</string>
-	<key>bisque</key>
-	<string>#ffe4c4</string>
-	<key>black</key>
-	<string>#000000</string>
-	<key>blanchedalmond</key>
-	<string>#ffebcd</string>
-	<key>blue</key>
-	<string>#0000ff</string>
-	<key>blueviolet</key>
-	<string>#8a2be2</string>
-	<key>brown</key>
-	<string>#a52a2a</string>
-	<key>burlywood</key>
-	<string>#deb887</string>
-	<key>cadetblue</key>
-	<string>#5f9ea0</string>
-	<key>chartreuse</key>
-	<string>#7fff00</string>
-	<key>chocolate</key>
-	<string>#d2691e</string>
-	<key>coral</key>
-	<string>#ff7f50</string>
-	<key>cornflowerblue</key>
-	<string>#6495ed</string>
-	<key>cornsilk</key>
-	<string>#fff8dc</string>
-	<key>crimson</key>
-	<string>#dc143c</string>
-	<key>cyan</key>
-	<string>#00ffff</string>
-	<key>darkblue</key>
-	<string>#00008b</string>
-	<key>darkcyan</key>
-	<string>#008b8b</string>
-	<key>darkgoldenrod</key>
-	<string>#b8860b</string>
-	<key>darkgray</key>
-	<string>#a9a9a9</string>
-	<key>darkgreen</key>
-	<string>#006400</string>
-	<key>darkgrey</key>
-	<string>#a9a9a9</string>
-	<key>darkkhaki</key>
-	<string>#bdb76b</string>
-	<key>darkmagenta</key>
-	<string>#8b008b</string>
-	<key>darkolivegreen</key>
-	<string>#556b2f</string>
-	<key>darkorange</key>
-	<string>#ff8c00</string>
-	<key>darkorchid</key>
-	<string>#9932cc</string>
-	<key>darkred</key>
-	<string>#8b0000</string>
-	<key>darksalmon</key>
-	<string>#e9967a</string>
-	<key>darkseagreen</key>
-	<string>#8fbc8f</string>
-	<key>darkslateblue</key>
-	<string>#483d8b</string>
-	<key>darkslategray</key>
-	<string>#2f4f4f</string>
-	<key>darkslategrey</key>
-	<string>#2f4f4f</string>
-	<key>darkturquoise</key>
-	<string>#00ced1</string>
-	<key>darkviolet</key>
-	<string>#9400d3</string>
-	<key>deeppink</key>
-	<string>#ff1493</string>
-	<key>deepskyblue</key>
-	<string>#00bfff</string>
-	<key>dimgray</key>
-	<string>#696969</string>
-	<key>dimgrey</key>
-	<string>#696969</string>
-	<key>dodgerblue</key>
-	<string>#1e90ff</string>
-	<key>firebrick</key>
-	<string>#b22222</string>
-	<key>floralwhite</key>
-	<string>#fffaf0</string>
-	<key>forestgreen</key>
-	<string>#228b22</string>
-	<key>fuchsia</key>
-	<string>#ff00ff</string>
-	<key>gainsboro</key>
-	<string>#dcdcdc</string>
-	<key>ghostwhite</key>
-	<string>#f8f8ff</string>
-	<key>gold</key>
-	<string>#ffd700</string>
-	<key>goldenrod</key>
-	<string>#daa520</string>
-	<key>gray</key>
-	<string>#808080</string>
-	<key>green</key>
-	<string>#008000</string>
-	<key>greenyellow</key>
-	<string>#adff2f</string>
-	<key>grey</key>
-	<string>#808080</string>
-	<key>honeydew</key>
-	<string>#f0fff0</string>
-	<key>hotpink</key>
-	<string>#ff69b4</string>
-	<key>indianred</key>
-	<string>#cd5c5c</string>
-	<key>indigo</key>
-	<string>#4b0082</string>
-	<key>ivory</key>
-	<string>#fffff0</string>
-	<key>khaki</key>
-	<string>#f0e68c</string>
-	<key>lavender</key>
-	<string>#e6e6fa</string>
-	<key>lavenderblush</key>
-	<string>#fff0f5</string>
-	<key>lawngreen</key>
-	<string>#7cfc00</string>
-	<key>lemonchiffon</key>
-	<string>#fffacd</string>
-	<key>lightblue</key>
-	<string>#add8e6</string>
-	<key>lightcoral</key>
-	<string>#f08080</string>
-	<key>lightcyan</key>
-	<string>#e0ffff</string>
-	<key>lightgoldenrodyellow</key>
-	<string>#fafad2</string>
-	<key>lightgray</key>
-	<string>#d3d3d3</string>
-	<key>lightgreen</key>
-	<string>#90ee90</string>
-	<key>lightgrey</key>
-	<string>#d3d3d3</string>
-	<key>lightpink</key>
-	<string>#ffb6c1</string>
-	<key>lightsalmon</key>
-	<string>#ffa07a</string>
-	<key>lightseagreen</key>
-	<string>#20b2aa</string>
-	<key>lightskyblue</key>
-	<string>#87cefa</string>
-	<key>lightslategray</key>
-	<string>#778899</string>
-	<key>lightslategrey</key>
-	<string>#778899</string>
-	<key>lightsteelblue</key>
-	<string>#b0c4de</string>
-	<key>lightyellow</key>
-	<string>#ffffe0</string>
-	<key>lime</key>
-	<string>#00ff00</string>
-	<key>limegreen</key>
-	<string>#32cd32</string>
-	<key>linen</key>
-	<string>#faf0e6</string>
-	<key>magenta</key>
-	<string>#ff00ff</string>
-	<key>maroon</key>
-	<string>#800000</string>
-	<key>mediumaquamarine</key>
-	<string>#66cdaa</string>
-	<key>mediumblue</key>
-	<string>#0000cd</string>
-	<key>mediumorchid</key>
-	<string>#ba55d3</string>
-	<key>mediumpurple</key>
-	<string>#9370db</string>
-	<key>mediumseagreen</key>
-	<string>#3cb371</string>
-	<key>mediumslateblue</key>
-	<string>#7b68ee</string>
-	<key>mediumspringgreen</key>
-	<string>#00fa9a</string>
-	<key>mediumturquoise</key>
-	<string>#48d1cc</string>
-	<key>mediumvioletred</key>
-	<string>#c71585</string>
-	<key>midnightblue</key>
-	<string>#191970</string>
-	<key>mintcream</key>
-	<string>#f5fffa</string>
-	<key>mistyrose</key>
-	<string>#ffe4e1</string>
-	<key>moccasin</key>
-	<string>#ffe4b5</string>
-	<key>navajowhite</key>
-	<string>#ffdead</string>
-	<key>navy</key>
-	<string>#000080</string>
-	<key>oldlace</key>
-	<string>#fdf5e6</string>
-	<key>olive</key>
-	<string>#808000</string>
-	<key>olivedrab</key>
-	<string>#6b8e23</string>
-	<key>orange</key>
-	<string>#ffa500</string>
-	<key>orangered</key>
-	<string>#ff4500</string>
-	<key>orchid</key>
-	<string>#da70d6</string>
-	<key>palegoldenrod</key>
-	<string>#eee8aa</string>
-	<key>palegreen</key>
-	<string>#98fb98</string>
-	<key>paleturquoise</key>
-	<string>#afeeee</string>
-	<key>palevioletred</key>
-	<string>#db7093</string>
-	<key>papayawhip</key>
-	<string>#ffefd5</string>
-	<key>peachpuff</key>
-	<string>#ffdab9</string>
-	<key>peru</key>
-	<string>#cd853f</string>
-	<key>pink</key>
-	<string>#ffc0cb</string>
-	<key>plum</key>
-	<string>#dda0dd</string>
-	<key>powderblue</key>
-	<string>#b0e0e6</string>
-	<key>purple</key>
-	<string>#800080</string>
-	<key>rebeccapurple</key>
-	<string>#663399</string>
-	<key>red</key>
-	<string>#ff0000</string>
-	<key>rosybrown</key>
-	<string>#bc8f8f</string>
-	<key>royalblue</key>
-	<string>#4169e1</string>
-	<key>saddlebrown</key>
-	<string>#8b4513</string>
-	<key>salmon</key>
-	<string>#fa8072</string>
-	<key>sandybrown</key>
-	<string>#f4a460</string>
-	<key>seagreen</key>
-	<string>#2e8b57</string>
-	<key>seashell</key>
-	<string>#fff5ee</string>
-	<key>sienna</key>
-	<string>#a0522d</string>
-	<key>silver</key>
-	<string>#c0c0c0</string>
-	<key>skyblue</key>
-	<string>#87ceeb</string>
-	<key>slateblue</key>
-	<string>#6a5acd</string>
-	<key>slategray</key>
-	<string>#708090</string>
-	<key>slategrey</key>
-	<string>#708090</string>
-	<key>snow</key>
-	<string>#fffafa</string>
-	<key>springgreen</key>
-	<string>#00ff7f</string>
-	<key>steelblue</key>
-	<string>#4682b4</string>
-	<key>tan</key>
-	<string>#d2b48c</string>
-	<key>teal</key>
-	<string>#008080</string>
-	<key>thistle</key>
-	<string>#d8bfd8</string>
-	<key>tomato</key>
-	<string>#ff6347</string>
-	<key>turquoise</key>
-	<string>#40e0d0</string>
-	<key>violet</key>
-	<string>#ee82ee</string>
-	<key>wheat</key>
-	<string>#f5deb3</string>
-	<key>white</key>
-	<string>#ffffff</string>
-	<key>whitesmoke</key>
-	<string>#f5f5f5</string>
-	<key>yellow</key>
-	<string>#ffff00</string>
-	<key>yellowgreen</key>
-	<string>#9acd32</string>
-</dict>
-</plist>
